requirements.txt
flask==3.0.3
python-binance==1.0.19
requests==2.31.0
